+++ fooof/synth/__init__.py
@@ -1,0 +1,3 @@
+from .gen import gen_freqs, gen_power_spectrum, gen_group_power_spectra
+from .params import Stepper, param_iter, param_sampler
+from .transform import rotate_spectrum
