+++ fooof/synth/gen.py
@@ -1,155 +1,14 @@
 """Synthesis functions for generating model components and synthetic power spectra."""
 
-from inspect import isgenerator
-from collections import namedtuple
-from itertools import chain, repeat
-
 import numpy as np
 
 from fooof.core.utils import group_three
 from fooof.core.funcs import gaussian_function, get_bg_func, infer_bg_func
+from fooof.synth.utils import _check_iter, _check_flat
+from fooof.synth.params import SynParams
 
 ###################################################################################################
 ###################################################################################################
-
-SynParams = namedtuple('SynParams', ['background_params', 'gaussian_params', 'nlv'])
-
-class Stepper():
-    """Object for stepping across parameter values.
-
-    Parameters
-    ----------
-    start : float
-        Start value to iterate from.
-    stop : float
-        End value to iterate to.
-    step : float
-        Incremet of each iteration.
-
-    Attributes
-    ----------
-    len : int
-        Length of generator range.
-    data : iterator
-        Set of specified parameters to iterate across.
-    """
-
-    def __init__(self, start, stop, step):
-
-        self._check_values(start, stop, step)
-
-        self.start = start
-        self.stop = stop
-        self.step = step
-        self.len = int((stop-start)/step)
-        self.data = iter(np.arange(start, stop, step))
-
-
-    def __len__(self):
-
-        return self.len
-
-
-    def __next__(self):
-
-        return round(next(self.data), 4)
-
-
-    def __iter__(self):
-
-        return self.data
-
-
-    @staticmethod
-    def _check_values(start, stop, step):
-        """Checks if provided values are valid for use."""
-
-        if any(i < 0 for i in [start, stop, step]):
-            raise ValueError("'start', 'stop', and 'step' should all be positive values")
-
-        if not start < stop:
-            raise ValueError("'start' should be less than stop")
-
-        if not step < (stop - start):
-            raise ValueError("'step' is too large given 'start' and 'stop' values")
-
-
-def param_iter(params):
-    """Generates parameters to iterate over.
-
-    Parameters
-    ----------
-    params : list of floats and Stepper
-        Parameters over which to iterate, where:
-            Stepper object defines iterated parameter and its range and,
-            floats are other parameters that will be held constant.
-
-    Yields
-    ------
-    list of floats
-        Next generated list of parameters.
-
-    Examples
-    --------
-    Iterates over center frequency values from 8 to 12 in increments of .25.
-
-    >>> osc = param_iter([Stepper(8, 12, .25), 1, 1])
-    """
-
-    # If input is a list of lists, check each element, and flatten if needed
-    if isinstance(params[0], list):
-        params = [item for sublist in params for item in sublist]
-
-    # Finds where Stepper object is. If there is more than one, raise an error
-    iter_ind = 0
-    num_iters = 0
-    for cur_ind, param in enumerate(params):
-        if isinstance(param, Stepper):
-            num_iters += 1
-            iter_ind = cur_ind
-
-        if num_iters > 1:
-            raise ValueError("Iteration is only supported on one parameter")
-
-    # Generate and yield next set of parameters
-    gen = params[iter_ind]
-    while True:
-        try:
-            params[iter_ind] = next(gen)
-            yield params
-        except StopIteration:
-            return
-
-
-def param_sampler(params, probs=None):
-    """Makes a generator to sample randomly from possible parameters.
-
-    Parameters
-    ----------
-    params : list of lists or list of float
-        Possible parameter values.
-    probs : list of float, optional
-        Probabilities with which to sample each parameter option. Default: None
-            If None, each parameter option is sampled uniformly.
-
-    Yields
-    ------
-    obj
-        A randomly sampled element from params.
-    """
-
-    # If input is a list of lists, check each element, and flatten if needed
-    if isinstance(params[0], list):
-        params = [_check_flat(lst) for lst in params]
-
-    # In order to use numpy's choice, with probabilities, choices are made on indices
-    # This is because the params can be a messy-sized list, that numpy choice does not like
-    inds = np.array(range(len(params)))
-
-    # While loop allows the generator to be called an arbitrary number of times.
-    while True:
-        yield params[np.random.choice(inds, p=probs)]
-
 
 def gen_freqs(freq_range, freq_res):
     """Generate a frequency vector, from the frequency range and resolution.
@@ -375,112 +234,3 @@ def gen_power_vals(xs, background_params, gauss_params, nlv):
     ys = np.power(10, background + peaks + noise)
 
     return ys
-
-
-def rotate_spectrum(freqs, power_spectrum, delta_f, f_rotation):
-    """Rotate a power spectrum about a frequency point, changing the power law exponent.
-
-    Parameters
-    ----------
-    freqs : 1d array
-        Frequency axis of input power spectrum, in Hz.
-    power_spectrum : 1d array
-        Power values of the spectrum that is to be rotated.
-    delta_f : float
-        Change in power law exponent to be applied.
-            Positive is counterclockwise rotation (flatten)
-            Negative is clockwise rotation (steepen).
-    f_rotation : float
-        Frequency, in Hz, at which to do the rotation, such that power at that frequency is unchanged.
-
-    Returns
-    -------
-    rotated_spectrum : 1d array
-        Rotated power spectrum.
-    """
-
-    # Check that the requested frequency rotation value is within the given range
-    if f_rotation < freqs.min() or f_rotation > freqs.max():
-        raise ValueError('Rotation frequency not within frequency range.')
-
-    f_mask = np.zeros_like(freqs)
-
-    f_mask = 10**(np.log10(np.abs(freqs)) * (delta_f))
-
-    # If starting freq is 0Hz, default power at 0Hz to keep same value because log will return inf.
-    if freqs[0] == 0.:
-        f_mask[0] = 1.
-
-    f_mask = f_mask / f_mask[np.where(freqs >= f_rotation)[0][0]]
-
-    rotated_spectrum = f_mask * power_spectrum
-
-    return rotated_spectrum
-
-###################################################################################################
-###################################################################################################
-
-def _check_iter(obj, length):
-    """Check an object to ensure that it is iterable, and make it iterable if not.
-
-    Parameters
-    ----------
-    obj : generator or list or float
-        Object to check status of.
-    length : int
-        The (minimum) length the iterator needs to be.
-
-    Returns
-    -------
-    obj : generator
-        Iterable object.
-    """
-
-    # If object is not a generator, update to become one
-    #   Otherwise (if the object already is a generator) then it gets left as it is
-    if not isgenerator(obj):
-
-        if isinstance(obj, list):
-
-            # Check if it's an empty list, if so make it a repeat generator of empty lists
-            if len(obj) == 0:
-                obj = repeat(obj)
-
-            # If obj is a list of lists of the right length, then we will leave it as is:
-            #   as a list of list that will iterate through each element
-            # If it is not, then it's turned into a repeat generator
-            # Note: checks that it's a list to not have an implicit error
-            #   when it's a list of numbers, that happens to be same length as n_spectra
-            elif not (isinstance(obj[0], list) and len(obj) == length):
-                obj = repeat(obj)
-
-        # If it's not a list, make it a repeat object (repeat int/float)
-        else:
-            obj = repeat(obj)
-
-    return obj
-
-
-def _check_flat(lst):
-    """Check whether a list is flat, and flatten if not.
-
-    Parameters
-    ----------
-    lst : list or list of lists
-        A list object to be checked and potentially flattened.
-
-    Returns
-    -------
-    list
-        A '1D' list, which is a flattened version of the input.
-
-    Notes
-    -----
-    This function only deals with one level of nesting.
-    """
-
-    # Note: flatten if list contains list(s), but skip if list is empty (which is valid)
-    if len(lst) != 0 and isinstance(lst[0], list):
-        lst = list(chain(*lst))
-
-    return lst
